+++ src/JWT/Action/FetchGooglePublicKeys/WithPsr6Cache.php
@@ -36,6 +36,7 @@ final class WithPsr6Cache implements Handler
 
         /** @noinspection PhpUnhandledExceptionInspection */
         $cacheItem = $this->cache->getItem($cacheKey);
+
         /** @var Keys|Expirable|null $keys */
         $keys = $cacheItem->get();